brian-mvp-backend/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
mysqlclient